+++ example/vite/src/main.tsx
@@ -1,28 +1,26 @@
+import "./index.css";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { GrazProvider, WalletType } from "graz";
-import { cosmoshub, osmosis } from "graz/chains";
+import { GrazProvider } from "graz";
+import { cosmoshub } from "graz/chains";
 import * as React from "react";
 import * as ReactDOM from "react-dom/client";
+
 import App from "./App";
-import "@getpara/react-sdk/styles.css";
-import { Environment } from "@getpara/react-sdk";
 
 const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
+  <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <GrazProvider
         grazOptions={{
-          chains: [cosmoshub, osmosis],
-          defaultWallet: WalletType.KEPLR,
-          paraConfig: {
-            apiKey: "3e185167932bc6cc7ed11c0abed4dfb8",
-            env:Environment.BETA
-          }
+          chains: [cosmoshub],
+
         }}
       >
         <App />
       </GrazProvider>
     </QueryClientProvider>
+  </React.StrictMode>,
 );