+++ packages/graz/src/actions/wallet/cosmostation.ts
@@ -39,7 +39,8 @@ export const getCosmostation = (): Wallet => {
       subscription,
       setDefaultOptions,
     });
-    return res;
+
+    return res as unknown as Wallet;
   }
 
   useGrazInternalStore.getState()._notFoundFn();