+++ packages/graz/src/actions/wallet/wallet-connect/index.ts
@@ -4,7 +4,6 @@ import type { Keplr } from "@keplr-wallet/types";
 import { SignClient } from "@walletconnect/sign-client";
 import type { ISignClient, SignClientTypes } from "@walletconnect/types";
 import { getSdkError } from "@walletconnect/utils";
-import Long from "long";
 
 import { useGrazInternalStore, useGrazSessionStore } from "../../../store";
 import { Key, type SignAminoParams, type SignDirectParams, type Wallet, WalletType } from "../../../types/wallet";
@@ -363,7 +362,7 @@ export const getWalletConnect = (params?: GetWalletConnectParams): Wallet => {
     return {
       signed: {
         chainId: signed.chainId ?? "",
-        accountNumber: signed.accountNumber ? Long.fromString(signed.accountNumber) : new Long(0),
+        accountNumber: signed.accountNumber ? BigInt(signed.accountNumber) : BigInt(0),
         authInfoBytes: signed.authInfoBytes
           ? new Uint8Array(Buffer.from(signed.authInfoBytes, encoding))
           : new Uint8Array([]),