+++ packages/graz/CHANGELOG.md
@@ -1,5 +1,57 @@
 # graz
 
+## 0.3.2
+
+### Patch Changes
+
+- ec1a66c: fix install deps warning
+
+## 0.3.1
+
+### Patch Changes
+
+- fff247b: fix type missmatch
+
+## 0.3.0
+
+### Minor Changes
+
+- df4ed52: move cosmjs deps to peerDeps
+
+### Patch Changes
+
+- 0525070: remove @cosmjs/launchpad package deps
+- 406b721: add @cosmjs/encoding as peer dependencies
+- 25b5d66: bump zustand v5
+- ab7afd7: bump walletconnect deps
+- 2fb86fd: bump cosmjs package
+- 72fb04d: remove terndermint client
+- 50e5e55: remove capsule
+
+## 0.2.6
+
+### Patch Changes
+
+- cd65f2f: This PR enable user to update wallet that connected with wallet connect
+
+## 0.2.5
+
+### Patch Changes
+
+- fa5d0fd: Fixes multi chain connect on non-snaps/dapp browser wallets - bug introduced in 0.2.3
+
+## 0.2.4
+
+### Patch Changes
+
+- 9240d0d: fixes build error
+
+## 0.2.3
+
+### Patch Changes
+
+- 1de0f39: We have updated from `getKey` to `getKeys` for accounts fetching in leap dapp browser as that is much faster.
+
 ## 0.2.1
 
 ### Patch Changes